Changed round to roundf

+++ Teensy/src/gpio/gpio.cpp
@@ -139,7 +139,7 @@ static int32_t calcWheelRev(volatile int *wheelCount)
     // Converts m/us to MPH
     float wheelSpeed = MS_TO_MPH(rps * wheelCircumference);
     // Converts it to an integer
-    int32_t wheelSpeedInt = static_cast<int32_t>(round(wheelSpeed));
+    int32_t wheelSpeedInt = static_cast<int32_t>(roundf(wheelSpeed));
     *wheelCount = 0;
     if (wheelSpeedInt < 0) {
         return static_cast<int32_t>(0);
